Add submodule for sonata-system
The automotive demo runs on both CHERIoT and non-CHERIoT, and to support
the build flow for this, requires access to some of the definitions in
`sonata-system` to be able to compile for a legacy baremetal
environment. This commit adds a submodule for `sonata-system` so that
these files can be used, without code duplication.