Add font support to the LCD `draw_str` libcall.
Extends the existing Sonata LCD library to support drawing a string with
a specified font, rather than always just using the m3x6 16pt font. The
original method is retained with that font as a default, but an
additional method is added via overloading which accepts an additional
font argument. To avoid exposing the internal font implementation, an
external-facing Font enum class is created which is then just mapped to
these internal fonts.

The LucidaConsole 10pt and 12pt fonts that are already in the display
drivers are then added as optional fonts that can be exposed.