Add receiving (board 2) automotive demo firmware
Adds example firmware for the receiving side of the automotive demo,
which will be running on the 2nd board. This requires the ethernet
driver, as well as the common automotive library.

The receiving firmware is responsible for receiving ethernet frames with
a specific structure over a connection. It has two possible modes of
operation, depending on which demo application is being run. For most
applications, it uses passthrough mode, which directly uses the incoming
acceleration pedal value to drive the car's motor. For specifically the
digital pedal app, it uses simulation mode, which uses the accelerator
pedal value as the driving force input to a model Eulerian physics
simulation, where the car's speed as it is driven against resistive
forces is loosely modeled via some arbitrary constants. The car's
simulated speed is then used to drive the car's motor instead.

The joystick can be pressed down to reset the car's speed between demos.