Add sending (board 1) auto demo CHERI firmware
Implements the base sending firmware (board 1) for the automotive demo,
running in CHERI on top of CHERIoT RTOS. Implements the common
automotive demo library, passing in callbacks with appropriate driver
functionality so that the automotive demos function on the sending board
side. This commit does not include support for the analogue pedal demo,
which requires use of the ADC.

The sending board is responsible for displaying a menu allowing the user
to select a demo application, and then either simulating a pedal or
polling a pedal (depending on the app selected), and transmitting this
pedal information the the receiving board (board 2) via an Ethernet
connection. It is also responsible for simultaneously running another
separate task that is bugged, such that it accidentally overwrites the
acceleration pedal value at some point. Linker script sections are used
to place these tasks contiguously in memory, albeit CHERI would not
allow the access without a capability violation either way.