Add initial automotive "no pedal" demo
This commit adds the initial common resources that are used by different
demonstration applications in the automotive demo, and adds the first
such application, which is the "no pedal" demo. This is a demo without
any user input, where a bug in a program that counts up and writes to an
array using that count over time causes an out-of-bounds array write
that overwrites some constant acceleration being transmitted.