Add automotive demo "digital pedal" application
This commit adds the "digital pedal" application to the automotive demo
library, which is an application that is very similar to the "Joystick
Pedal" demo in its operation, but that instead uses the digital joystick
read callback implemented by users to read the value of some actual
external pedal as a digital input. The demo is initialised in simulation
mode to allow a reasonable demo application with a digital input, and
the same bugged behaviour as in the Joystick pedal demo is incorporated,
with the user moving the joystick to manually trigger the bug, or
pressing the joystick to end the demo.