Add automotive demo "analogue pedal" application
This commit adds the main "Analogue Pedal" application that will be used
to showcase the demo, relying on the implementing firmware having access
to a working ADC. The analogue pedal read callback is used to read the
input analogue value coming from some external pedal, and this is then
passed through to the receiving board via Ethernet.

Unlike with the other 3 demo applications added so far, this application
features a more visual and realistic bugged task, which is a small
"volume slider" application which allows the user to move the joystick
left or right to increase or decrease the volume correspondingly.
However, a purposeful bug in the code allows the volume to increase one
past its maximum value, to 21/20. The volume value is used every frame
to calculate colour information that is placed inside a small
"framebuffer" array to make a nice coloured volume bar slider, except
when the bugged behaviour occurs, this colour will overwrite the value
of the analogue accelerator pedal that has been read, causing a
wrong / very large acceleration to be transmitted.