Add simple legacy ADC driver
Implements a simple legacy (X)ADC driver for use in the automotive demo.
As the XADC is already configured with defaults such that it starts by
sampling all connected analogue channels and system measurements, this
driver simply provides an easily-extensible interface for reading the
latest ADC measurements from the mapped XADC DRP registers, and for
using the configuration registers to modify the divider / clock speed to
potentially reduce power draw if your application requires this, and
does not need to use the ADC. This is based on the upstream CHERIoT RTOS
CHERIoT version of this driver.