Add sending (board 1) auto demo legacy firmware
Implements the base sending firmware (board 1) for the automotive demo,
running in legacy mode (non-CHERI), baremetal. Implements the common
automotive demo library, passing in callbacks with appropriate driver /
system functionality so that the automotive demo functions on the
sending board side. This commit does not include support for the
analogue pedal demo, which requires use of the ADC.

This commit also includes a simple build flow for legacy baremetal
firmware in xmake based on the existing cmake flow in the sonata-system
submodule, allowing cross-compilation for RISC-V non-CHERI. As part of
this, the rv32imcb GCC compiler toolchain is added to `flake.nix`.