Fix nix linting of submodules
Linting was breaking on including `sonata-system` as a submodule, due to
one of the vendored files from coremark in `sonata-system` containing a
non-UTF-8 copyright symbol in one of its markdown files. This broke
lychee linting.

In general, it doesn't make sense to lint submodules included under the
`third_party` directory. The exclude path contained `vendor`, which does
not exist, but not `third_party` - this is likely an error from copying
or a typo. This commit excludes that path from the lychee lint to avoid
these lints, fixing errors.

mypy was also running into python module naming conflicts with a
`test_runner` in both `sonata-software` and `sonata-system`, so a flag
has similarly been added to the `mypy` linting call to exclude the
`third_party` directory.