Add Rego primer & firmware auditing exercise 1
This commit introduces exercise 1 for using CHERIoT Audit to audit
firmware alongside Rego policies. This exercise involves writing a
policy to detect if firmware uses sealed capabilities, and to query
information about these capabilities if they exist.

This also includes a quick primer on Rego, with an introduction to
some of the language's features.