Replace Nix `env-with-everything` with `build-legacy-software`
Removes the `env-with-everything` environment, as there isn't much
practical reason to have an environment where you build the simulator
and also have the legacy toolchain, exclusive to Linux.

Instead, it makes more sense to retain the existing simulator
environment, and add a separate environment `build-legacy-software` that
can be used to get the legacy toolchain on Linux systems. This means
that entering this environment for building legacy software is much
quicker, and the same functionality can be retained by overlaying the
two environments (entering the sim environment, then the legacy one).