Update demos to use updated I2C driver.
The updated I2C driver handles controller halts properly, and for good
practice requires (via a `nodiscard` attribute) that the new Boolean
result of the I2C's `blocking_write` function is not discarded. This
commit hence simply adds checks to the existing demos that will debug
log if an error occurs on these writes.