Update LCD library to use LCD CS and PWM instead of GPIO
In the latest 1.0 hardware, the LCD has been changed to use a chip
select (for its CS, DC and RST pins) and to use a PWM for its backlight.
This commit updates the LCD library/driver to appropriately make use of
the SPI driver intsead of the GPIO as it was using before, and to use
the PWM for the LCD's backlight. The PWM is always set to 100, and for
now the driver simply replaces existing functionality - there is no way
to modify the PWM from the driver.