Update Examples/Exercises to use updated GPIO/PWM drivers
Sonata's updated GPIO driver correctly provides different struct for
different GPIO devices. This updates Sonata's examples and exercises
(with the exception of the automotive demo, which will be handled
separately) to use the new drivers and names of devices from the board
description file, and changes direct reads/writes to/from the `input`
and `output` registers to match the new general/board GPIO mappings
after things have moved in the hardware.