Update guide for using updated pinmux driver interface
Sonata's Pinmux driver has been updated to give it more granaular
capabilites, such that capabilities to individual pin sinks and block
sinks can be created and distributed, for better compartmentalisation
and isolation of software that uses pinmux.

This commit updates `sonata-software` to use the updated driver.
Currently, the only thing using Pinmux in the software repo is the IO
guide, and so this is updated to reflect the new software interface for
this driver.