Heartbleed: Fix regression in legacy demo from LCD updates
While the legacy automotive component was updated and the other demos
were updated, the legacy Heartbleed demo was only partially updated -
but the demo was failing to build. This was not caught in CI due to the
fact that, unlike the legacy automotive demo, the legacy heartbleed demo
build is not yet included in CI at the moment.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>