nix: Add heartbleed legacy build to CI quality check
Unlike the automotive demo's legacy component, the heartbleed demo's
legacy component had not been added to the Nix flake and the
corresponding CI .yml file, meaning that regressions in the ability to
build the software were not being caught. Add this so that future
breakages to the demo will also cause CI errors.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>